removed trash & too complicated image-conditioned example
