+++ examples/mnist_gans/configs/vanilla_gan.yml
@@ -1,29 +1,60 @@
-common_hyperparams:
-  noise_dim: &noise_dim 16
-  # ensure the names below are the same as in your runner ("discriminator_train" & "generator_train")
-  discriminator_train_phase: &discriminator_train_phase discriminator_train
+runner_params:
+  # input keys
+  # Note: for albumentations transforms we have to have key "image" =(
+  data_input_key: &real_data "image"
+  class_input_key: &class_targets "class_targets"
+  noise_input_key: &noise_input "noise"
+  # output keys
+  fake_logits_output_key: &fake_logits "fake_logits"
+  real_logits_output_key: &real_logits "real_logits"
+  fake_data_output_key: &fake_data "fake_image"
+  # phases
   generator_train_phase: &generator_train_phase generator_train
+  discriminator_train_phase: &discriminator_train_phase discriminator_train
+  # model keys:
+  generator_model_key: &generator_model_name "generator"
+  discriminator_model_key: &discriminator_model_name "discriminator"
 
 model_params:
   _key_value: True
   generator:
-    model: SimpleCImageGenerator
-    noise_dim: *noise_dim
+    model: SimpleGenerator
+    noise_dim: &noise_dim 16
   discriminator:
-    model: SimpleCImageDiscriminator
+    model: SimpleDiscriminator
 
 
 args:
-  expdir: "mnist_advanced_gans"
-  baselogdir: "./logs/mnist_advanced_gans/vanilla_cigan_onehot"
+  expdir: "mnist_gans"
+  baselogdir: "./logs/mnist_gans/vanilla_gan"
 
 
 stages:
 
+  transform_params:
+    transform: Compose
+    transforms:
+      - transform: AsImage
+      - transform: Normalize
+        mean: [0.5]
+        std: [0.5]
+      - transform: ToTensorV2
+      - transform: AdditionalNoiseTensor
+        tensor_size: [*noise_dim]
+        output_key: *noise_input
+      - transform: AdditionalScalar
+        value: 1.
+        output_key: &real_targets "real_targets"
+      - transform: AdditionalScalar
+        value: 0.
+        output_key: &fake_targets "fake_targets"
+
   data_params:
     batch_size: 64
     num_workers: 0
-    use_same_class_batch_sampler: True  # special sampler to generate batches of same classes (batch["targets"][:len(batch) // 2] == batch["targets"][len(batch) // 2:])
+
+    image_key: *real_data
+    target_key: *class_targets
 
   state_params:
     num_epochs: 100
@@ -43,26 +74,13 @@ stages:
         *discriminator_train_phase: 1
         *generator_train_phase: 1
 
-    prepare_batch_same_class:
-      callback: SameClassFeaturesRepeatCallback
-      data_key: "data"
-      same_class_data_key: "same_class_data"
-      targets_key: "class_targets"  # set to none (remove string) to speed up
-
-    prepare_batch_noise:
-      callback: AddBatchNoiseCallback
-      noise_shape: *noise_dim
-
-    prepare_batch_targets:
-      callback: AddRealFakeTargetsCallback
-
     loss_g:
       _wrapper: &g_train_wrapper
         callback: PhaseBatchWrapperCallback
         active_phases: [*generator_train_phase]
       callback: CriterionCallback
-      input_key: "real_targets"
-      output_key: "fake_logits"
+      input_key: *real_targets
+      output_key: *fake_logits
       prefix: loss_g
 
     loss_d_real:
@@ -70,14 +88,14 @@ stages:
         callback: PhaseBatchWrapperCallback
         active_phases: [*discriminator_train_phase]
       callback: CriterionCallback
-      input_key: "real_targets"
-      output_key: "real_logits"
+      input_key: *real_targets
+      output_key: *real_logits
       prefix: loss_d_real
     loss_d_fake:
       _wrapper: *d_train_wrapper
       callback: CriterionCallback
-      input_key: "fake_targets"
-      output_key: "fake_logits"
+      input_key: *fake_targets
+      output_key: *fake_logits
       prefix: loss_d_fake
     loss_d:
       _wrapper: *d_train_wrapper
@@ -99,7 +117,7 @@ stages:
 
     visualizer:
       callback: VisualizationCallback
-      output_keys: "fake_data"
+      output_keys: *fake_data
       n_row: 5
       max_images: 25
 
@@ -112,9 +130,9 @@ stages:
       _key_value: True
       generator:
         optimizer: Adam
-        _model: ["generator"]
+        _model: [*generator_model_name]
         lr: 0.0002
       discriminator:
         optimizer: Adam
-        _model: ["discriminator"]
+        _model: [*discriminator_model_name]
         lr: 0.0002
