+++ examples/mnist_gans/configs/wasserstein_gan.yml
@@ -28,8 +28,8 @@ model_params:
 
 
 args:
-  expdir: "mnist_advanced_gans"
-  baselogdir: "./logs/mnist_advanced_gans/vanilla_wgan_gp"
+  expdir: "mnist_gans"
+  baselogdir: "./logs/mnist_gans/wasserstein_gan"
 
 
 stages:
@@ -60,11 +60,7 @@ stages:
     batch_consistant_metrics: False
 
   criterion_params:
-    _key_value: True
-    mean_output_loss:
-      criterion: MeanOutputLoss
-    gradient_penalty:
-      criterion: GradientPenaltyLoss
+    criterion: MeanOutputLoss
 
   callbacks_params:
     phase_manager:
@@ -82,7 +78,6 @@ stages:
       callback: CriterionCallback
       input_key: *dummy_input_key  # input key does not matter
       output_key: *fake_validity
-      criterion_key: mean_output_loss
       prefix: loss_g
       multiplier: -1.0
 
@@ -93,23 +88,13 @@ stages:
       callback: CriterionCallback
       input_key: *dummy_input_key  # input key does not matter
       output_key: *real_validity
-      criterion_key: mean_output_loss
       prefix: loss_d_real
     loss_d_fake:
       _wrapper: *d_train_wrapper
       callback: CriterionCallback
       input_key: *dummy_input_key  # input key does not matter
       output_key: *fake_validity
-      criterion_key: mean_output_loss
       prefix: loss_d_fake
-    loss_d_gp:
-      _wrapper: *d_train_wrapper
-      callback: CriterionWithDiscriminatorCallback
-      input_key: *real_data
-      output_key: *fake_data
-      criterion_key: gradient_penalty
-      discriminator_model_key: *critic_model_name
-      prefix: "loss_d_gp"
     loss_d:
       _wrapper: *d_train_wrapper
       callback: CriterionAggregatorCallback
@@ -118,7 +103,6 @@ stages:
       loss_keys:
         loss_d_real: -1.0
         loss_d_fake: 1.0
-        loss_d_gp: 10.0  # gradient penalty multiplier
 
     optim_g:
       _wrapper: *g_train_wrapper
@@ -127,9 +111,10 @@ stages:
       loss_key: loss_g
     optim_d:
       _wrapper: *d_train_wrapper
-      callback: OptimizerCallback
+      callback: WeightClampingOptimizerCallback
       optimizer_key: discriminator
       loss_key: loss_d
+      weight_clamp_value: 0.01
 
     wasserstein_distance:
       _wrapper: *d_train_wrapper
@@ -154,14 +139,8 @@ stages:
       generator:
         optimizer: Adam
         _model: [*generator_model_name]
-        lr: 0.0001
-        betas:
-          - 0.5
-          - 0.9
+        lr: 0.00005
       discriminator:
         optimizer: Adam
         _model: [*critic_model_name]
-        lr: 0.0001
-        betas:
-          - 0.5
-          - 0.9
+        lr: 0.00005
