+++ examples/mnist_gans/criterion.py
