+++ examples/mnist_gans/models.py
@@ -98,34 +98,9 @@ class SimpleCDiscriminator(nn.Module):
         return x
 
 
-class SimpleCImageGenerator(SimpleGenerator):
-    def __init__(
-        self,
-        noise_dim=10,
-        hidden_dim=256,
-        image_resolution=(28, 28),
-        channels=1
-    ):
-        super().__init__(
-            noise_dim + np.prod(image_resolution) * channels, hidden_dim,
-            image_resolution, channels
-        )
-
-    def forward(self, z, c_image):
-        x = torch.cat((z, c_image.reshape(c_image.size(0), -1)), dim=1)
-        return super().forward(x)
-
-
-class SimpleCImageDiscriminator(SimpleDiscriminator):
-    def __init__(self, image_resolution=(28, 28), channels=1, hidden_dim=100):
-        super().__init__(image_resolution, channels * 2, hidden_dim)
-
-    def forward(self, x, c_image):
-        return super().forward(torch.cat((x, c_image), dim=1))
-
-
 __all__ = [
-    "SimpleGenerator", "SimpleDiscriminator", "SimpleCGenerator",
-    "SimpleCDiscriminator", "SimpleCImageDiscriminator",
-    "SimpleCImageGenerator"
+    "SimpleGenerator",
+    "SimpleDiscriminator",
+    "SimpleCGenerator",
+    "SimpleCDiscriminator"
 ]
