+++ examples/mnist_gans/runners.py
@@ -96,47 +96,3 @@ class CGanRunner(GanRunner):
             generator_model_key=generator_model_key,
             discriminator_model_key=discriminator_model_key
         )
-
-
-class ICGanRunner(CGanRunner):
-    """
-    (Image) Conditional GAN
-        both generator and discriminator are conditioned on same class image
-        assumed usage:
-            another_image_of_class_c = generator(noise, image_of_class_c)
-            real_score = discriminator(image1_class_c, image2_class_c)
-            fake_score = discriminator(gen_image_class_c, image2_class_c)
-    """
-    def __init__(
-        self,
-        model: Union[Model, Dict[str, Model]] = None,
-        device: Device = None,
-        data_input_key: str = "data",
-        class_input_key: str = "class_targets",
-        noise_input_key: str = "noise",
-        fake_logits_output_key: str = "fake_logits",
-        real_logits_output_key: str = "real_logits",
-        fake_data_output_key: str = "fake_data",
-        d_fake_condition_key: str = "same_class_data",
-        d_real_condition_key: str = "same_class_data",
-        generator_train_phase: str = "generator_train",
-        discriminator_train_phase: str = "discriminator_train",
-        generator_model_key: str = "generator",
-        discriminator_model_key: str = "discriminator"
-    ):
-        super().__init__(
-            model=model,
-            device=device,
-            data_input_key=data_input_key,
-            class_input_key=class_input_key,
-            noise_input_key=noise_input_key,
-            fake_logits_output_key=fake_logits_output_key,
-            real_logits_output_key=real_logits_output_key,
-            fake_data_output_key=fake_data_output_key,
-            d_fake_condition_key=d_fake_condition_key,
-            d_real_condition_key=d_real_condition_key,
-            generator_train_phase=generator_train_phase,
-            discriminator_train_phase=discriminator_train_phase,
-            generator_model_key=generator_model_key,
-            discriminator_model_key=discriminator_model_key
-        )
