+++ examples/mnist_gans/transforms.py
